Add files via upload
This file contains a set of 50 data points (see file cars.csv). Each data point is a triplet, where the first component is the index, the second component is the speed, and the third component is distance. The observations give the speed of cars and the distances taken to stop.